feat(userApi.js, Post.jsx): add functionality to mark/unmark a post as favorite
Add two new functions to the userApi.js file to add and delete a post from the user's favorite list. In the Post.jsx file, add a new button to mark/unmark a post as favorite. When the user clicks on the button, the markPostAsFavorite function is called, which sends a request to the server to add or delete the post from the user's favorite list. The button's icon changes depending on whether the post is marked as a favorite or not.