feat(CommentForm.jsx): add functionality to create a comment and update parent component with new comment
The changes in this commit add functionality to the CommentForm component. The component now includes a handleSubmit function that creates a new comment object with the content entered by the user. The comment object is then passed to the createComment function from the postApi module, which makes an API call to create the comment. If the API call is successful, the setCommentCreated function is called to update the parent component with the new comment. If there is an error, it is logged to the console. Finally, the content state is reset to an empty string.